+++ src/main/kotlin/com/github/nvelychenko/drupalextend/extensions/PsiElementExtensions.kt
@@ -1,4 +1,4 @@
-package com.github.nvelychenko.drupalextend.util
+package com.github.nvelychenko.drupalextend.extensions
 
 import com.intellij.psi.PsiElement