+++ src/main/kotlin/com/github/nvelychenko/drupalextend/type/StaticContentEntityTypeProvider.kt
@@ -1,7 +1,7 @@
 package com.github.nvelychenko.drupalextend.type
 
 import com.github.nvelychenko.drupalextend.index.ContentEntityFqnIndex
-import com.github.nvelychenko.drupalextend.util.getValue
+import com.github.nvelychenko.drupalextend.extensions.getValue
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
@@ -18,6 +18,14 @@ class StaticContentEntityTypeProvider : PhpTypeProvider4 {
         return '\u0434'
     }
 
+    private val splitKey = '\u0435'
+
+    private val possibleMethods = mutableMapOf(
+        Pair("load", ""),
+        Pair("create", "[]"),
+        Pair("loadMultiple", "[]"),
+    )
+
     override fun getType(psiElement: PsiElement): PhpType? {
         if (DumbService.getInstance(psiElement.project).isDumb) {
             return null
@@ -25,16 +33,16 @@ class StaticContentEntityTypeProvider : PhpTypeProvider4 {
 
         if (psiElement !is MethodReference) return null
 
-        if (psiElement.name != "load" || !psiElement.isStatic) return null
+        if (!possibleMethods.containsKey(psiElement.name)) return null
 
-        return PhpType().add("#$key${psiElement.signature}")
+        return PhpType().add("#$key${psiElement.signature}$splitKey${psiElement.name}")
     }
 
     override fun complete(expression: String?, project: Project?): PhpType? {
         if (expression == null || project == null || !expression.contains(key))
             return null
 
-        val signature = expression.substring(2)
+        val (signature, methodName) = expression.replace("#$key", "").split(splitKey)
 
         val fileBasedIndex = FileBasedIndex.getInstance()
 
@@ -42,7 +50,8 @@ class StaticContentEntityTypeProvider : PhpTypeProvider4 {
             if (partialSignature.startsWith("#M#C")) {
                 val contentEntity = fileBasedIndex.getValue(ContentEntityFqnIndex.KEY, partialSignature.substring(4).substringBefore('.'), project) ?: continue
 
-                return PhpType().add(contentEntity.fqn)
+                val methodType = possibleMethods[methodName] ?: ""
+                return PhpType().add(contentEntity.fqn + methodType)
             }
         }