+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # drupal-extend Changelog
 
+## [Unreleased]
+
+### Added
+
+- Integration with symfony service type provider for \Drupal::service
+
 ## [0.0.5]
 
 ### Added
@@ -16,3 +22,6 @@
 - Fields autocomplete was not working in certain circumstances
 - Content entity provider was not working in certain circumstances
 - Deprecated usage of baseDir
+
+[Unreleased]: https://github.com/nvelychenko/drupal-extend/compare/0.0.5...HEAD
+[0.0.5]: https://github.com/nvelychenko/drupal-extend/commits/0.0.5